# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Management Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get projects",
                "operationId": "9590eef2a9081e6591e1ce49c8360ec1",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Load related resources (comma separated)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "creator,company,tasks,files,comments,members"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get projects successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "digikala-api"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "digikala is the largest online store in Iran."
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1998-07-28 00:56:01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2009-08-26 23:51:20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Project"
                ],
                "summary": "Project create",
                "operationId": "98061d39b8df330fe77ccb0b1a20d708",
                "requestBody": {
                    "description": "Project data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "project name",
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "digikala-api"
                                    },
                                    "description": {
                                        "description": "description for project",
                                        "maxLength": 400,
                                        "example": "digikala is the largest online store in Iran.",
                                        "nullable": true
                                    },
                                    "members": {
                                        "description": "members id for project",
                                        "type": "string",
                                        "example": "[1,2,3,4,5]"
                                    },
                                    "start_date": {
                                        "description": "project date start",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-07-28 00:56:01"
                                    },
                                    "end_date": {
                                        "description": "project date start",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-09-28 00:56:01"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "create project successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "digikala-api"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "digikala is the largest online store in Iran."
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1998-07-28 00:56:01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2009-08-26 23:51:20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/projects/{project}": {
            "get": {
                "tags": [
                    "Project"
                ],
                "summary": "Get project",
                "operationId": "06bd1e61a9a76aa02f4d7a06212e352b",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Load related resources (comma separated)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "creator,company,tasks,files,comments,members"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get project successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "digikala-api"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "digikala is the largest online store in Iran."
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1998-07-28 00:56:01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2009-08-26 23:51:20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Project"
                ],
                "summary": "Project update",
                "operationId": "342c493b23cb55f7057cfe2210dc2d8f",
                "requestBody": {
                    "description": "Project update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "project name",
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "digikala-api",
                                        "nullable": true
                                    },
                                    "description": {
                                        "description": "description for project",
                                        "type": "string",
                                        "maxLength": 400,
                                        "example": "digikala is the largest online store in Iran.",
                                        "nullable": true
                                    },
                                    "start_date": {
                                        "description": "project date start",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-07-28 00:56:01",
                                        "nullable": true
                                    },
                                    "end_date": {
                                        "description": "project date start",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-09-28 00:56:01",
                                        "nullable": true
                                    },
                                    "members": {
                                        "description": "update members id for project",
                                        "example": "[1,2,3,4,5,6,7,8]",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": "1",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "update project successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "digikala-api"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "digikala is the largest online store in Iran."
                                                },
                                                "start_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "1998-07-28 00:56:01"
                                                },
                                                "end_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2009-08-26 23:51:20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Project"
                ],
                "summary": "delete a project",
                "operationId": "04cc8f2a338f8a37fc7bd5adf4f0508b",
                "responses": {
                    "204": {
                        "description": "project delete successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}